+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/FloatGeneratorTest.kt
@@ -19,13 +19,16 @@ class FloatGeneratorTest : AbstractGeneratorTest() {
 
     @Test
     fun produceInsideGivenRange() {
-        assertTrue(Generator.floats(-8f, 14f).randoms(42).take(200).all { it >= -8 && it <= 14 })
+        assertTrue(Generator.floats(-8f, 14f).randoms(0).take(1000).all { it >= -8 && it <= 14 })
     }
 
     @Test
-    fun startsWithEdgeCases() {
-        val edgeCases = Generator.floats().randoms(42).take(3).toSet()
+    fun providesSamples() {
+        assertEquals(setOf(0f, -1f, 1f, -Float.MAX_VALUE, Float.MAX_VALUE), Generator.floats().samples)
+    }
 
-        assertEquals(setOf(0f, -1f, 1f), edgeCases)
+    @Test
+    fun samplesAreInRange() {
+        assertEquals(setOf(1f, Float.MAX_VALUE), Generator.floats(min = 1f).samples)
     }
 }