+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/MapGeneratorTest.kt
@@ -19,13 +19,23 @@ class MapGeneratorTest : AbstractGeneratorTest() {
     }
 
     @Test
-    fun emitsEmptyLists() {
-        assertTrue(Generator.maps<Int, Double>().randoms(0).take(200).any { it.isEmpty() })
+    fun samplesContainsEmpty() {
+        assertTrue(Generator.maps<Int, Double>().samples.any { it.isEmpty() })
     }
 
     @Test
-    fun emitsSingletonsLists() {
-        assertTrue(Generator.maps<Int, Double>().randoms(0).take(200).any { it.size == 1 })
+    fun samplesContainsSingletons() {
+        assertTrue(Generator.maps<Int, Double>().samples.any { it.size == 1 })
+    }
+
+    @Test
+    fun noEmptySampleWhenMinSizeIsGreaterThan0() {
+        assertTrue(Generator.maps(Generator.ints(), Generator.doubles(), minSize = 1).samples.none { it.isEmpty() })
+    }
+
+    @Test
+    fun noSingletonSampleWhenMinSizeIsGreaterThan1() {
+        assertTrue(Generator.maps(Generator.ints(), Generator.doubles(), minSize = 2).samples.none { it.size <= 1 })
     }
 
     @Test