+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/operator/FilterTest.kt
@@ -1,9 +1,6 @@
 package com.github.jcornaz.kwik.operator
 
-import com.github.jcornaz.kwik.AbstractGeneratorTest
-import com.github.jcornaz.kwik.Generator
-import com.github.jcornaz.kwik.filter
-import com.github.jcornaz.kwik.filterNot
+import com.github.jcornaz.kwik.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
@@ -19,6 +16,15 @@ class FilterTest : AbstractGeneratorTest() {
             assertEquals(0, it % 2)
         }
     }
+
+    @Test
+    fun filterSamples() {
+        val generator = Generator.create { it.nextInt() }
+            .withSamples(42, 77)
+            .filter { it < 50 }
+
+        assertEquals(setOf(42), generator.samples)
+    }
 }
 
 class FilterNotTest : AbstractGeneratorTest() {
@@ -31,5 +37,14 @@ class FilterNotTest : AbstractGeneratorTest() {
         gen.randoms(0).take(1000).forEach {
             assertNotEquals(0, it % 2)
         }
+    }
+
+    @Test
+    fun filterSamples() {
+        val generator = Generator.create { it.nextInt() }
+            .withSamples(42, 77)
+            .filterNot { it < 50 }
+
+        assertEquals(setOf(77), generator.samples)
     }
 }