+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/operator/MapTest.kt
@@ -3,11 +3,12 @@ package com.github.jcornaz.kwik.operator
 import com.github.jcornaz.kwik.AbstractGeneratorTest
 import com.github.jcornaz.kwik.Generator
 import com.github.jcornaz.kwik.map
+import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class MapTest : AbstractGeneratorTest() {
-    override val generator: Generator<*> = Generator.create { it.nextInt() }.map { it.toString() }
+    override val generator: Generator<String> = Generator.create { it.nextInt() }.map { it.toString() }
 
     @Test
     fun applyTransform() {
@@ -27,5 +28,14 @@ class MapTest : AbstractGeneratorTest() {
         val transformedResults = transformed.randoms(37).take(200).toSet()
 
         assertEquals(sourceResults.size, transformedResults.size)
+    }
+
+    @Test
+    fun transformSamples() {
+        val generator = Generator.create { it.nextInt() }
+            .withSamples(42, 77)
+            .map { it.toString() }
+
+        assertEquals(setOf("42", "77"), generator.samples)
     }
 }