+++ CHANGELOG.rst
@@ -16,6 +16,23 @@ Added
 .....
 
 * ``filter`` operator for generators
+* **(Breaking)** ``samples`` property in the ``Generator`` interface to improve management of the value samples (edge-cases)
+* ``forAll`` non-inline function, allowing to hide implementation details.
+
+Changed
+.......
+
+* **(Breaking)** Renamed arguments ``from`` and ``until`` of ``floats`` and ``doubles`` generators
+  for more consistency with the int and long generators
+* **(Breaking)** Make the lambda of ``forAll`` crossinline, to allow usage of a non-inline function and hide implementation details.
+* Now the size probability for strings an collections generators is similar for all sizes.
+  instead they have empty and singletons instances as samples
+* ``combine`` and ``combineWith`` now start by a combination of the source generators and will randomly add samples in the random generations.
+
+Removed
+.......
+
+* **(Breaking)** ``ratio`` argument from the ``withSamples`` and ``withNull`` operators
 
 0.1.0-alpha.3_ - 2019-06-08
 ---------------------------